+++ src/main/resources/static/bookNode.js
